examples: pull resource scoping infra into benchmark
